# docs/api/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "카페 프로젝트 명세서",
    "version": "0.0.1",
    "title": "Cafe Project API",
    "contact": {
      "name": "김용환",
      "url": "https://yonghwankim-dev.tistory.com/",
      "email": "yh.kim951107@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/qna/{id}": {
      "get": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 조회",
        "operationId": "detailQuestionUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "댓글 커서",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "게시글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 수정",
        "operationId": "modifyQuestionUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "수정 게시글 등록번호",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionRequest",
            "description": "수정 게시글 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionSavedRequest"
            }
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보가 담긴 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "질문 게시글 API 정보를 제공하는 Controller"
        ],
        "summary": "질문 삭제",
        "operationId": "deleteQuestionUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "삭제 게시글 등록번호",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보가 담긴 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/QuestionResponse"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/qna/{id}/comments": {
      "get": {
        "tags": [
          "댓글 API 정보를 제공하는 Controller"
        ],
        "summary": "댓글 목록 조회",
        "operationId": "listCommentUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "댓글 커서",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionId",
            "description": "게시글 등록번호",
            "required": false
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "댓글 API 정보를 제공하는 Controller"
        ],
        "summary": "댓글 추가",
        "operationId": "createCommentUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "commentRequest",
            "description": "댓글 추가 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentSavedRequest"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "questionId",
            "description": "게시글 등록번호",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommentResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/qna/{id}/comments/{commentId}": {
      "put": {
        "tags": [
          "댓글 API 정보를 제공하는 Controller"
        ],
        "summary": "댓글 수정",
        "operationId": "modifyCommentUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "query",
            "description": "댓글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "commentRequest",
            "description": "댓글 수정 정보",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentSavedRequest"
            }
          },
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionId",
            "description": "게시글 등록번호",
            "required": false
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommentResponse"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "댓글 API 정보를 제공하는 Controller"
        ],
        "summary": "댓글 삭제",
        "operationId": "deleteCommentUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "query",
            "description": "댓글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionId",
            "description": "게시글 등록번호",
            "required": false
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CommentResponse"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/qna/{id}/comments/{commentId}/edit": {
      "get": {
        "tags": [
          "댓글 API 정보를 제공하는 Controller"
        ],
        "summary": "댓글 수정 페이지",
        "operationId": "editCommentFormUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "query",
            "description": "댓글 등록번호",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "creationTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "id",
            "in": "query",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastAccessedTime",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "maxInactiveInterval",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "new",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "questionId",
            "description": "게시글 등록번호",
            "required": false
          },
          {
            "in": "query",
            "name": "session",
            "description": "로그인 정보 세션",
            "required": false
          },
          {
            "name": "valueNames",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ModelAndView"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CommentResponse": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "questionId": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "writerName": {
          "type": "string"
        }
      },
      "title": "CommentResponse"
    },
    "CommentSavedRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "minLength": 1,
          "maxLength": 3000
        },
        "questionId": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "CommentSavedRequest"
    },
    "ModelAndView": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "boolean"
        },
        "model": {
          "type": "object"
        },
        "modelMap": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "reference": {
          "type": "boolean"
        },
        "status": {
          "type": "string",
          "enum": [
            "ACCEPTED",
            "ALREADY_REPORTED",
            "BAD_GATEWAY",
            "BAD_REQUEST",
            "BANDWIDTH_LIMIT_EXCEEDED",
            "CHECKPOINT",
            "CONFLICT",
            "CONTINUE",
            "CREATED",
            "DESTINATION_LOCKED",
            "EXPECTATION_FAILED",
            "FAILED_DEPENDENCY",
            "FORBIDDEN",
            "FOUND",
            "GATEWAY_TIMEOUT",
            "GONE",
            "HTTP_VERSION_NOT_SUPPORTED",
            "IM_USED",
            "INSUFFICIENT_SPACE_ON_RESOURCE",
            "INSUFFICIENT_STORAGE",
            "INTERNAL_SERVER_ERROR",
            "I_AM_A_TEAPOT",
            "LENGTH_REQUIRED",
            "LOCKED",
            "LOOP_DETECTED",
            "METHOD_FAILURE",
            "METHOD_NOT_ALLOWED",
            "MOVED_PERMANENTLY",
            "MOVED_TEMPORARILY",
            "MULTIPLE_CHOICES",
            "MULTI_STATUS",
            "NETWORK_AUTHENTICATION_REQUIRED",
            "NON_AUTHORITATIVE_INFORMATION",
            "NOT_ACCEPTABLE",
            "NOT_EXTENDED",
            "NOT_FOUND",
            "NOT_IMPLEMENTED",
            "NOT_MODIFIED",
            "NO_CONTENT",
            "OK",
            "PARTIAL_CONTENT",
            "PAYLOAD_TOO_LARGE",
            "PAYMENT_REQUIRED",
            "PERMANENT_REDIRECT",
            "PRECONDITION_FAILED",
            "PRECONDITION_REQUIRED",
            "PROCESSING",
            "PROXY_AUTHENTICATION_REQUIRED",
            "REQUESTED_RANGE_NOT_SATISFIABLE",
            "REQUEST_ENTITY_TOO_LARGE",
            "REQUEST_HEADER_FIELDS_TOO_LARGE",
            "REQUEST_TIMEOUT",
            "REQUEST_URI_TOO_LONG",
            "RESET_CONTENT",
            "SEE_OTHER",
            "SERVICE_UNAVAILABLE",
            "SWITCHING_PROTOCOLS",
            "TEMPORARY_REDIRECT",
            "TOO_EARLY",
            "TOO_MANY_REQUESTS",
            "UNAUTHORIZED",
            "UNAVAILABLE_FOR_LEGAL_REASONS",
            "UNPROCESSABLE_ENTITY",
            "UNSUPPORTED_MEDIA_TYPE",
            "UPGRADE_REQUIRED",
            "URI_TOO_LONG",
            "USE_PROXY",
            "VARIANT_ALSO_NEGOTIATES"
          ]
        },
        "view": {
          "$ref": "#/definitions/View"
        },
        "viewName": {
          "type": "string"
        }
      },
      "title": "ModelAndView"
    },
    "QuestionResponse": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "writer": {
          "type": "string"
        }
      },
      "title": "QuestionResponse"
    },
    "QuestionSavedRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "title": {
          "type": "string",
          "pattern": "^.{1,100}$"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "QuestionSavedRequest"
    },
...
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}